pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel>=0.44.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-tc-generator"
version = "1.2.0"
description = "Context-Rich Test Case Generator with YAML Prompt Management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AI TC Generator Team"}
]
keywords = ["test-generation", "ai", "reqif", "automotive", "yaml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
requires-python = ">=3.13.7"
dependencies = [
    "pandas>=2.3.0,<2.4.0",
    "requests>=2.32.0,<2.33.0", 
    "PyYAML>=6.0.2,<6.1.0",
    "click>=8.2.0,<8.3.0",
    "rich>=13.9.0,<14.0.0",
    "openpyxl>=3.1.5,<3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "mypy>=1.16.0,<1.17.0",
    "ruff>=0.8.0,<1.0.0",
    "types-requests>=2.32.0,<3.0.0",
    "types-PyYAML>=6.0.0,<7.0.0",
]
security = [
    "pip-audit>=2.9.0,<3.0.0",
    "safety>=3.6.0,<4.0.0",
    "pip-licenses>=4.4.0,<5.0.0",
]
performance = [
    "lxml>=5.3.0,<6.0.0",
    "orjson>=3.10.0,<4.0.0",
    "uvloop>=0.21.0,<1.0.0; sys_platform != 'win32'",
]
profiling = [
    "memory-profiler>=0.61.0,<1.0.0",
    "py-spy>=0.4.0,<1.0.0",
]

[project.scripts]
ai-tc-generator = "src.generate_contextual_tests_v002:main"

[project.urls]
Homepage = "https://github.com/your-org/AI_TC_Generator_v03_Excel"
Repository = "https://github.com/your-org/AI_TC_Generator_v03_Excel"
Issues = "https://github.com/your-org/AI_TC_Generator_v03_Excel/issues"
Documentation = "https://github.com/your-org/AI_TC_Generator_v03_Excel/blob/main/README.md"

# ============================================================================
# TOOL CONFIGURATIONS
# ============================================================================

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "utilities*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.md"]

# ============================================================================
# RUFF CONFIGURATION (Linting and Formatting)
# ============================================================================

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings  
    "F",      # Pyflakes
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "I",      # isort
    "N",      # pep8-naming
    "C4",     # flake8-comprehensions
    "PTH",    # flake8-use-pathlib
    "T20",    # flake8-print
    "RET",    # flake8-return
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "PL",     # pylint
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value used in comparison
    "T201",   # print statements (allowed for CLI output)
]

[tool.ruff.lint.per-file-ignores]
"utilities/version_check.py" = ["T201"]  # Allow prints in version checker
"src/generate_contextual_tests_v002.py" = ["T201"]  # Allow prints in main app

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============================================================================
# MYPY CONFIGURATION (Type Checking)
# ============================================================================

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

# Module-specific overrides
[[tool.mypy.overrides]]
module = [
    "yaml",
    "requests",
    "pandas", 
    "openpyxl",
    "rich",
    "click",
]
ignore_missing_imports = true

# ============================================================================
# PYTEST CONFIGURATION (Testing)
# ============================================================================

[tool.pytest.ini_options]
minversion = "8.4"
addopts = [
    "--strict-config",
    "--strict-markers", 
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests", 
    "unit: marks tests as unit tests",
    "performance: marks tests as performance benchmarks",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

# ============================================================================
# COVERAGE CONFIGURATION
# ============================================================================

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]